+++ src/pages/knowledgegraph/GraphSearch/GraphSearch.tsx
@@ -1,34 +1,14 @@
-import { useEffect, useMemo, useRef, useState } from 'react';
-import { useTranslation } from 'react-i18next';
+import { useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
-import { W3CAnnotation } from '@annotorious/react';
+import { useTranslation } from 'react-i18next';
+import { Grip, X } from 'lucide-react';
 import { useDraggable } from '@neodrag/react';
-import { CirclePlus, Grip, PanelsTopLeft, Trash2, X } from 'lucide-react';
-import { useOpenInAnnotationView } from '@/pages/annotate/AnnotationViewState';
+import { W3CAnnotation } from '@annotorious/react';
 import { Button } from '@/ui/Button';
-import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/Tooltip';
-import { useSearchDialogPos, useSearchState } from '../KnowledgeGraphState';
-import { GraphSearchConditionBuilder } from './GraphSearchConditionBuilder';
-import { ExportSelector } from './export';
-import { useFulltextSearch } from './useFulltextSearch';
-import { 
-  Condition, 
-  Graph, 
-  GraphNode, 
-  GraphNodeType, 
-  KnowledgeGraphSettings,
-  NestedConditionSentence,
-  Operator, 
-  Sentence
-} from '../Types';
-import { 
-  Select, 
-  SelectContent, 
-  SelectItem,
-  SelectTrigger, 
-  SelectValue 
-} from '@/ui/Select';
-import { Input } from '@/ui/Input';
+import { Graph, GraphNode, KnowledgeGraphSettings } from '../Types';
+import { useSearchDialogPos } from '../KnowledgeGraphState';
+import { FulltextSearch } from './fulltext/FullTextSearch';
+import { Builder } from './builder';
 
 interface GraphSearchProps {
 
@@ -48,8 +28,6 @@ interface GraphSearchProps {
 
 }
 
-const EMPTY_CONDITION: Condition = { operator: 'AND', sentence: {} };
-
 export const GraphSearch = (props: GraphSearchProps) => {
 
   const { t } = useTranslation('knowledgegraph');
@@ -58,119 +36,12 @@ export const GraphSearch = (props: GraphSearchProps) => {
 
   const { position, setPosition } = useSearchDialogPos({ x: props.isFullscreen ? 10 : 260, y: 10 });
 
-  const { objectType, setObjectType, conditions, setConditions } = useSearchState();
-
-  const { search } = useFulltextSearch(props.annotations);
-
-  const { openInAnnotationView } = useOpenInAnnotationView();
-
-  const matchedImages = useMemo(() => {
-    if (objectType !== 'IMAGE' || !props.query) return [];
-    
-    return props.graph.nodes
-      .filter(n => props.query!(n))
-      .map(m => m.id);
-  }, [props.query, props.graph, objectType]);
-
+  const [tab, setTab] = useState<'BUILDER' | 'FULLTEXT'>('BUILDER');
+  
   useDraggable(el, {
     position,
     onDrag: ({ offsetX, offsetY }) => setPosition({ x: offsetX, y: offsetY })
   });
-
-  useEffect(() => {
-    if (conditions.length === 0) {
-      // No conditions - remove query
-      props.onChangeQuery(undefined)
-    } else {
-      // Remove the last condition if it is unfinished
-      const toApply = conditions[conditions.length - 1].matches ?
-        conditions : conditions.slice(0, -1);
-
-      // For now, we'll treat the condition list step by step, where
-      // all conditions BEFORE the current are treated as if they were one
-      // result. E.g: 
-      //
-      //  'A' and 'B' and 'C' or 'D'
-      //
-      //  would be logically treated as...
-      // 
-      //  ((('A' and 'B') and 'C') or 'D')
-      const matches = new Set(toApply.reduce<string[]>((previousMatches, condition) => {
-        if (condition.operator === 'AND') {
-          // Next result is the intersection of the previous with this result
-          return previousMatches.filter(str => (condition.matches || []).includes(str));
-        } else {
-          // Next result is the union of the previous with this result
-          return [...new Set([...previousMatches, ...(condition.matches || [])])];
-        }        
-      }, conditions[0].matches!));
-
-      const query = (n: GraphNode) =>
-        n.type === objectType && matches.has(n.id);
-
-      props.onChangeQuery(query);
-    }
-  }, [conditions]);
-
-  const isComplete = (sentence: Partial<Sentence>) => {
-    if (!sentence.ConditionType) return false;
-
-    if ('Attribute' in sentence && sentence.Attribute) {
-      // SimpleConditionSentence
-      return Boolean(sentence.Comparator);
-    } else if ('data' in sentence) {
-      return Boolean(sentence.data);
-    } else {
-      // NestedConditionSentence
-      return Boolean((sentence as NestedConditionSentence).Value);
-    }
-  }
-  
-  const onChange = (current: Partial<Sentence>, next: Partial<Sentence>, matches?: string[]) => {
-    const updated = conditions.map(condition => 
-        condition.sentence === current ? ({ operator: condition.operator, sentence : next, matches }) : condition);
-
-    setConditions(updated);
-  }
-
-  const onDelete = (sentence: Partial<Sentence>) => {
-    const next = conditions.filter(c => c.sentence !== sentence);
-
-    setConditions(next);
-
-    if (next.length === 0)
-      setObjectType(undefined);
-  }
-
-  const onSelectObjectType = (value: string) => {
-    setObjectType(value as GraphNodeType);
-
-    if (value) {
-      setConditions([{...EMPTY_CONDITION}]);
-    } else {
-      setConditions([]);
-    }
-  }
-
-  const onSelectOperator = (sentence: Partial<Sentence>, operator: Operator) => {
-    const next = conditions.map(c => c.sentence === sentence ? ({...c, operator }) : c);
-    setConditions(next);
-  }
-
-  const onClearAll = () => {
-    setObjectType(undefined);
-    setConditions([]);
-  }
-
-  const [query, setQuery] = useState('');
-
-  useEffect(() => {
-    const hits = search(query);
-    if (!hits) return; 
-
-    const q = (n: GraphNode) => new Set((hits.map(h => h.nodeId))).has(n.id);
-    props.onChangeQuery(q);
-  }, [query, search]);
 
   return createPortal(
     <div 
@@ -193,136 +64,16 @@ export const GraphSearch = (props: GraphSearchProps) => {
       </div>
 
       <div className="px-3 pr-5 pb-2">
-        <Input 
-          value={query}
-          onChange={e => setQuery(e.target.value)}
-        />
-
-        <div className="text-xs flex items-center gap-2">
-          <span className="w-14 text-right">
-            {t('graphSearch.find')}
-          </span>
-          
-          <Select 
-            value={objectType || ''}
-            onValueChange={onSelectObjectType}>
-            <SelectTrigger className="rounded-none px-2 py-1 h-auto bg-white shadow-none">
-              <span className="text-xs">
-                <SelectValue placeholder={t('graphSearch.selectNodeType')} />
-              </span>
-            </SelectTrigger>
-
-            <SelectContent>
-              {props.settings.graphMode === 'RELATIONS' && (
-                <SelectItem
-                  className="text-xs" 
-                  value="ENTITY_TYPE">{t('graphSearch.objectTypes.entityClasses')}</SelectItem>
-              )}
-
-              {props.settings.includeFolders && (
-                <SelectItem
-                  className="text-xs" 
-                  value="FOLDER">{t('graphSearch.objectTypes.subFolders')}</SelectItem>
-              )}
-
-              <SelectItem
-                className="text-xs" 
-                value="IMAGE">{t('graphSearch.objectTypes.images')}</SelectItem>
-            </SelectContent>
-          </Select>
-        </div>
-
-        {conditions.map(({ operator, sentence }, idx) => (
-          <div 
-            className="flex flex-nowrap gap-2 pt-2 text-xs items-start"
-            key={idx}>
-            
-            {(idx === 0) ? (
-              <div className="w-14" />
-            ) : (
-              <div className="w-14">
-                <Select 
-                  value={operator || 'AND'}
-                  onValueChange={op => onSelectOperator(sentence, op as Operator)}>
-                  <SelectTrigger className="w-16 rounded-none border-r-0 px-2 py-1 h-auto bg-white shadow-none">
-                    <span className="text-xs">
-                      <SelectValue />
-                    </span>
-                  </SelectTrigger>
-
-                  <SelectContent>
-                    <SelectItem
-                      className="text-xs" 
-                      value="AND">{t('graphSearch.operators.and')}</SelectItem>
-
-                    <SelectItem
-                      className="text-xs"
-                      value="OR">{t('graphSearch.operators.or')}</SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-            )}
-
-            <GraphSearchConditionBuilder 
-              annotations={props.annotations}
-              graph={props.graph}
-              objectType={objectType}
-              sentence={sentence}
-              settings={props.settings}
-              onChange={(next, matches) => onChange(sentence, next, matches)}
-              onDelete={() => onDelete(sentence)} />
-          </div>
-        ))}
-
-        {(conditions.length > 0 && isComplete(conditions[conditions.length - 1].sentence)) ? (
-          <div className="flex justify-between pt-4 pl-14">
-            <div className="flex items-center gap-4">
-              <Button 
-                disabled={!conditions.map(c => c.sentence).every(isComplete)}
-                variant="link"
-                size="sm"
-                className="flex items-center text-xs py-0 px-0 font-normal"
-                onClick={() => setConditions(conditions => ([...conditions, {...EMPTY_CONDITION}]))}>
-                <CirclePlus className="h-3.5 w-3.5 ml-0.5 mr-1 mb-0.5" /> {t('graphSearch.addCondition')}
-              </Button>
-
-              <Button 
-                variant="link"
-                size="sm"
-                className="flex items-center text-xs py-0 px-0 font-normal"
-                onClick={onClearAll}>
-                <Trash2 className="h-3.5 w-3.5 mr-1 mb-px" /> {t('graphSearch.clearAll')}
-              </Button>
-            </div>
-
-            {props.query && (
-              <div className="flex items-center gap-2">
-                <ExportSelector 
-                  objectType={objectType}
-                  graph={props.graph} 
-                  query={props.query} />
-
-                <Tooltip>
-                  <TooltipTrigger asChild>
-                    <Button 
-                      disabled={matchedImages.length === 0}
-                      variant="ghost"
-                      size="sm"
-                      className="h-7 gap-1 text-xs font-normal px-1.5"
-                      onClick={() => openInAnnotationView(matchedImages)}>
-                      {matchedImages.length} <PanelsTopLeft className="size-3.5" />
-                    </Button>
-                  </TooltipTrigger>
-
-                  <TooltipContent>
-                    {t('graphSearch.openInWorkspace')}
-                  </TooltipContent>
-                </Tooltip>
-              </div>
-            )}
-          </div>
+        {tab === 'BUILDER' ? (
+          <Builder 
+            annotations={props.annotations}
+            graph={props.graph}
+            isFullscreen={props.isFullscreen}
+            settings={props.settings} 
+            query={props.query}
+            onChangeQuery={props.onChangeQuery} />
         ) : (
-          <div className="h-4"/>
+          <FulltextSearch />
         )}
       </div>
     </div>, document.body
