+++ src/pages/knowledgegraph/GraphSearch/builder/GraphSearchConditionBuilder.tsx
@@ -18,7 +18,7 @@ import {
   Sentence, 
   SimpleConditionSentence, 
   SubCondition
-} from '../Types';
+} from '../../Types';
 import { 
   Select, 
   SelectContent,
