+++ src/pages/knowledgegraph/GraphSearch/fulltext/FullTextSearch.tsx
@@ -1,0 +1,7 @@
+export const FulltextSearch = () => {
+
+  return (
+    <div>asdfasdf asdfasdf</div>
+  )
+  
+}
